+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 #### Changed
 - Implement flood-specific severity calculation for `storms.noaa` events.
+- Do not override event type for `storms.noaa` events in feed composition.
 
 #### Removed